+++ docs/api/changelog.rst
@@ -60,6 +60,12 @@ Version 0.10
   to coroutines, a callback function can exploit the strong ordering guarantee
   of the XMPP XML Stream.
 
+* Support for passing a callback function to
+  :meth:`aioxmpp.stream.StanzaStream.send` which is invoked on responses to an
+  IQ request sent through :meth:`~aioxmpp.stream.StanzaStream.send`. In contrast
+  to awaiting the result of :meth:`~aioxmpp.stream.StanzaStream.send`, the
+  callback can exploit the strong ordering guarantee of the XMPP XML Stream.
+
 .. _api-changelog-0.9:
 
 Version 0.9